backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.10.0
pinecone>=5.0.0
redis>=5.0.1
prisma>=0.11.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
tiktoken>=0.5.2
numpy>=1.26.3
